Fix cluster and kube controller sync (#38491)
* Split out cluster and clusterstore from secretcontroller

* Respect remote cluster timeout when running each cluster and kubecontroller

* Fix lint

* Fix case when provided a error secret, the callback failed, so the cluster has no chance to be run

* reduce max elapsed time

* Fix close twice

* Fix lint

* Remove SyncTimeout for local cluster

* Set timeout in kube multi cluster controller for non local clusters
